fix: supporting all means of death in the frag run pattern
add: UDT_cutter commands to add/remove means of death
chg: extended udtFragRunPatternArg::AllowedMeansOfDeaths to 64 bits
chg: smarter way to generate the udtPlayerMeanOfDeath list

+++ UDT_GUI/src/App.cs
@@ -78,7 +78,7 @@ namespace Uber.DemoTools
     {
         public Int32 PatternCutPlayerIndex = int.MinValue; // @NOTE: Some negative values have meaning already.
         public string PatternCutPlayerName = "";
-        public UInt32 FragCutAllowedMeansOfDeaths = 0;
+        public UInt64 FragCutAllowedMeansOfDeaths = 0;
         public UDT_DLL.udtProtocol ConversionOutputProtocol = UDT_DLL.udtProtocol.Invalid;
         public bool QuitAfterFirstJob = false;
         public bool ForceAnalyzeOnLoad = false;
